// server/api/spotify/types.ts
export interface SimplifiedArtist {
    external_urls: ExternalUrls;
    href: string;
    id: string;
    name: string;
    type: string;
    uri: string;
}

export interface Artist extends SimplifiedArtist {
    followers: Followers;
    genres: string[];
    images: Image[];
    popularity: number;
}

export interface Artists {
    artists: Artist[];
}

export interface FollowedArtists {
    artists: Page<Artist>;
}

export interface Followers {
    href: string | null;
    total: number;
}

export interface ExternalUrls {
    spotify: string;
}
// ...
export interface Image {
    url: string;
    height: number;
    width: number;
}